+++ src/slurm_plugin/clustermgtd.py
@@ -40,7 +40,7 @@ from common.time_utils import seconds
 from common.utils import check_command_output, read_json, sleep_remaining_loop_time, time_is_up, wait_remaining_time
 from retrying import retry
 from slurm_plugin.cluster_event_publisher import ClusterEventPublisher
-from slurm_plugin.common import TIMESTAMP_FORMAT, event_publisher, log_exception, print_with_count
+from slurm_plugin.common import TIMESTAMP_FORMAT, log_exception, print_with_count
 from slurm_plugin.console_logger import ConsoleLogger
 from slurm_plugin.instance_manager import InstanceManager
 from slurm_plugin.slurm_resources import (
@@ -398,10 +398,8 @@ class ClusterManager:
             self._task_executor = self._initialize_executor(config)
 
             self._config = config
-            self._event_publisher = ClusterEventPublisher(
-                event_publisher(
-                    event_logger, config.cluster_name, "HeadNode", "clustermgtd", config.head_node_instance_id
-                )
+            self._event_publisher = ClusterEventPublisher.create(
+                event_logger, config.cluster_name, "HeadNode", "clustermgtd", config.head_node_instance_id
             )
             self._compute_fleet_status_manager = ComputeFleetStatusManager()
             self._instance_manager = self._initialize_instance_manager(config)
