+++ src/slurm_plugin/resume.py
@@ -21,14 +21,12 @@ from botocore.config import Config
 from common.schedulers.slurm_commands import get_nodes_info, set_nodes_down
 from common.utils import read_json
 from slurm_plugin.cluster_event_publisher import ClusterEventPublisher
-from slurm_plugin.common import event_publisher, event_publisher_noop, is_clustermgtd_heartbeat_valid, print_with_count
+from slurm_plugin.common import is_clustermgtd_heartbeat_valid, print_with_count
 from slurm_plugin.instance_manager import InstanceManager
 from slurm_plugin.slurm_resources import CONFIG_FILE_DIR
 
 log = logging.getLogger(__name__)
 event_logger = log.getChild("events")
-
-_event_publisher: ClusterEventPublisher = ClusterEventPublisher(event_publisher_noop)
 
 
 class SlurmResumeConfig:
@@ -204,7 +202,15 @@ def _resume(arg_nodes, resume_config):
         )
         for error_code, node_list in instance_manager.failed_nodes.items():
             _handle_failed_nodes(node_list, reason=f"(Code:{error_code})Failure when resuming nodes")
-    _event_publisher.publish_node_launch_events(instance_manager.failed_nodes)
+
+        _event_publisher = ClusterEventPublisher.create(
+            event_logger,
+            resume_config.cluster_name,
+            "HeadNode",
+            "slurm-resume",
+            resume_config.head_node_instance_id,
+        )
+        _event_publisher.publish_node_launch_events(instance_manager.failed_nodes)
 
 
 def main():
@@ -231,17 +237,6 @@ def main():
                 default_log_file,
                 e,
             )
-        global _event_publisher
-        _event_publisher = ClusterEventPublisher(
-            event_publisher(
-                event_logger,
-                resume_config.cluster_name,
-                "HeadNode",
-                "slurm-resume",
-                resume_config.head_node_instance_id,
-            )
-        )
-
         log.info("ResumeProgram config: %s", resume_config)
         _resume(args.nodes, resume_config)
         log.info("ResumeProgram finished.")
