+++ tests/slurm_plugin/test_common.py
@@ -8,15 +8,13 @@
 # or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
-import json
 import logging
 from datetime import datetime, timedelta, timezone
-from types import SimpleNamespace
 
 import pytest
 from assertpy import assert_that
 from common.utils import read_json, time_is_up
-from slurm_plugin.common import TIMESTAMP_FORMAT, event_publisher, get_clustermgtd_heartbeat
+from slurm_plugin.common import TIMESTAMP_FORMAT, get_clustermgtd_heartbeat
 
 
 @pytest.mark.parametrize(
@@ -108,506 +106,3 @@ def test_read_json(test_datadir, caplog, json_file, default_value, raises_except
         assert_that(caplog.text).matches(message_in_log)
     else:
         assert_that(caplog.text).does_not_match("exception")
-
-
-@pytest.mark.parametrize(
-    "log_level, base_args, events, expected_events",
-    [
-        (
-            logging.INFO,
-            {"a-setting": "value-a"},
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "a-setting": "value-b",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                )
-            ],
-            [
-                {
-                    "a-setting": "value-b",
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                },
-            ],
-        ),
-        (
-            logging.INFO,
-            {"a-setting": "value-a"},
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "a-setting": "value-b",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                ),
-                (
-                    "DEBUG",
-                    "debug-message",
-                    "debug-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "a-setting": "value-b",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                ),
-            ],
-            [
-                {
-                    "a-setting": "value-b",
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                },
-            ],
-        ),
-        (
-            logging.INFO,
-            {"a-setting": "value-a", "b-setting": "global-b-setting"},
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "a-setting": "value-b",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                )
-            ],
-            [
-                {
-                    "a-setting": "value-b",
-                    "b-setting": "global-b-setting",
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                },
-            ],
-        ),
-        (
-            logging.INFO,
-            {"a-setting": "value-a"},
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "a-setting": "value-b",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                ),
-                (
-                    "DEBUG",
-                    "debug-message",
-                    "debug-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "a-setting": "value-b",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                ),
-                (
-                    "WARNING",
-                    "warning-message",
-                    "warning-event",
-                    {
-                        "datetime": "a-date-and-time",
-                        "detail": {
-                            "a-detail": "detail-a",
-                        },
-                    },
-                ),
-            ],
-            [
-                {
-                    "a-setting": "value-b",
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                },
-                {
-                    "a-setting": "value-a",
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "WARNING",
-                    "instance-id": "instance_id",
-                    "event-type": "warning-event",
-                    "message": "warning-message",
-                    "detail": {"a-detail": "detail-a"},
-                },
-            ],
-        ),
-    ],
-)
-def test_event_publisher(log_level, base_args, events, expected_events):
-    received_events = []
-
-    def log_handler(level, format_string, value):
-        received_events.append([level, value])
-
-    metric_logger = SimpleNamespace()
-    metric_logger.isEnabledFor = lambda level: level >= log_level
-    metric_logger.log = log_handler
-
-    publisher = event_publisher(metric_logger, "cluster", "HeadNode", "component", "instance_id", **base_args)
-
-    # Run test
-    for event in events:
-        publisher(event[0], event[1], event[2], **event[3])
-
-    # Assert calls
-    assert_that(received_events).is_length(len(expected_events))
-    for actual, expected in zip(received_events, expected_events):
-        assert_that(actual[0]).is_greater_than_or_equal_to(log_level)
-        actual_json = json.loads(actual[1])
-        assert_that(actual_json).is_equal_to(expected)
-
-
-@pytest.mark.parametrize(
-    "log_level, events, expected_events, expected_supplied_count",
-    [
-        (
-            logging.INFO,
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    [
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                    ],
-                ),
-            ],
-            [
-                {
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                    "a_setting": "value-b",
-                    "b_setting": "b-kwargs-setting",
-                },
-            ],
-            1,
-        ),
-        (
-            logging.INFO,
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    [
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                    ],
-                ),
-            ],
-            [
-                {
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                    "a_setting": "value-b",
-                    "b_setting": "b-kwargs-setting",
-                },
-                {
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                    "a_setting": "value-b",
-                    "b_setting": "b-kwargs-setting",
-                },
-            ],
-            2,
-        ),
-        (
-            logging.INFO,
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    [
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                    ],
-                ),
-                (
-                    "DEBUG",
-                    "info-message",
-                    "info-event",
-                    [
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "debug-value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "debug-value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                    ],
-                ),
-            ],
-            [
-                {
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                    "a_setting": "value-b",
-                    "b_setting": "b-kwargs-setting",
-                },
-                {
-                    "datetime": "a-date-and-time",
-                    "version": 0,
-                    "scheduler": "slurm",
-                    "cluster-name": "cluster",
-                    "node-role": "HeadNode",
-                    "component": "component",
-                    "level": "INFO",
-                    "instance-id": "instance_id",
-                    "event-type": "info-event",
-                    "message": "info-message",
-                    "detail": {"a-detail": "detail-a"},
-                    "a_setting": "value-b",
-                    "b_setting": "b-kwargs-setting",
-                },
-            ],
-            2,
-        ),
-        (
-            logging.WARNING,
-            [
-                (
-                    "INFO",
-                    "info-message",
-                    "info-event",
-                    [
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                    ],
-                ),
-                (
-                    "DEBUG",
-                    "info-message",
-                    "info-event",
-                    [
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "debug-value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                        {
-                            "datetime": "a-date-and-time",
-                            "a_setting": "debug-value-b",
-                            "detail": {
-                                "a-detail": "detail-a",
-                            },
-                        },
-                    ],
-                ),
-            ],
-            [],
-            0,
-        ),
-    ],
-)
-def test_event_publisher_with_event_supplier(log_level, events, expected_events, expected_supplied_count):
-    received_events = []
-    events_supplied = 0
-
-    def log_handler(level, format_string, value):
-        received_events.append(value)
-
-    def event_supplier(events_to_supply):
-        nonlocal events_supplied
-        for supplied_event in events_to_supply:
-            yield supplied_event
-            events_supplied += 1
-
-    metric_logger = SimpleNamespace()
-    metric_logger.isEnabledFor = lambda level: level >= log_level
-    metric_logger.log = log_handler
-
-    publisher = event_publisher(metric_logger, "cluster", "HeadNode", "component", "instance_id")
-
-    for event in events:
-        publisher(
-            event[0],
-            event[1],
-            event[2],
-            event_supplier=event_supplier(event[3]),
-            a_setting="a-kwargs-setting",
-            b_setting="b-kwargs-setting",
-        )
-
-    assert_that(events_supplied).is_equal_to(expected_supplied_count)
-
-    assert_that(received_events).is_length(len(expected_events))
-    for actual, expected in zip(received_events, expected_events):
-        actual_json = json.loads(actual)
-        assert_that(actual_json).is_equal_to(expected)
-
-
-def test_event_publisher_swallows_exceptions(caplog):
-    handler_called = False
-
-    def log_handler(level, format_string, value):
-        nonlocal handler_called
-        handler_called = True
-        raise Exception("hello")
-
-    metric_logger = SimpleNamespace()
-    metric_logger.isEnabledFor = lambda level: True
-    metric_logger.log = log_handler
-
-    publisher = event_publisher(metric_logger, "cluster", "HeadNode", "component", "instance_id")
-
-    publisher(logging.INFO, "hello", "event-type", detail={"hello": "goodbye"})
-
-    assert_that(handler_called).is_true()
-
-    assert_that(caplog.records).is_length(1)
